Change type int to short int

+++ zeabus_library_cpp/include/zeabus/fuzzy/control_error_3dimension.hpp
@@ -68,7 +68,7 @@ namespace fuzzy
 
             // this use set fuzzy rule for 3Dimension
             void set_fuzzy_rule( 
-                std::array< std::array < std::array < int, 7 > , 7 > , 7 >* ptr_fuzzy_rule );
+                std::array< std::array < std::array < short int, 7> , 7> , 7>* ptr_fuzzy_rule );
 
             // this use defuzzification rule
             void set_defuzzification_rule( std::array< double , 3 >* ptr_defuzzy_rule );
@@ -87,7 +87,7 @@ namespace fuzzy
             // force rule will have to 4 part because we want to get value for bound force
             std::array< double , 4 >* ptr_force_rule; // rule for fuzzification force range
             std::array< double , 3 >* ptr_defuzzy_rule; // rule for defuzzification relative 
-            std::array< std::array < std::array < int , 7 > , 7 > , 7 >* ptr_fuzzy_rule;
+            std::array< std::array < std::array < short int , 7 > , 7 > , 7 >* ptr_fuzzy_rule;
 
             // Use to publish data
             // This message_pub will use to calculate all data in fuzzy too
